Game Failed Window & Employee Tower Shop
Now, when players failed the game, the failed window will pop up to show the failed information with related UI function.
Tower Shop UI structure is ready, but I haven't figured out how to connect the tower UI element with the actual drag and set actions.
Basically, the Player UI the, FailedCanvas, and the TowerShopCanvas and related components are done.

Next Steps:
1. When click the restart button to reload the MainScene, the grid and target path line will disappear
2. Connect the shop UI with the actual  tower setting actions
3. Connect the Shop UI money changes with the towers' and enemies' actions (first need to discuss the rules with group members maybe)
4. Textures and effects implementation
5. Apply the in-class knowledge with the game elements